Not a voting recommendation, but what they promise
Not sure if this is the best phrasing, but the idea is to sound
like fightforthefuture's "saves the net"/"breaks the net" language.

IMHO, we don't want to get into a situation where we recommend
voting for a party (regardless of all they promise/do) according to
their opinion re a specific subject.